+++ setup.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-#!/bin/bash
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: MIT-0